requirements.txt
django 
djangorestframework 
djangorestframework-simplejwt
drf-yasg
drf-spectacular
langchain
langchain-huggingface
chromadb
python-dotenv
pydantic
langchain-community
pypdf
langchain-chroma
sentence-transformers
python-docx
apscheduler
docx2txt

